feat: #1501 - refactored user preferences files into a dedicated folders
