+++ packages/smooth_app/lib/database/product_query.dart
@@ -6,7 +6,7 @@ import 'package:smooth_app/data_models/product_list.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/database/dao_string.dart';
 import 'package:smooth_app/database/local_database.dart';
-import 'package:smooth_app/pages/user_preferences_dev_mode.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_dev_mode.dart';
 import 'package:uuid/uuid.dart';
 
 abstract class ProductQuery {
