+++ packages/smooth_app/lib/helpers/smooth_matched_product.dart
@@ -5,7 +5,7 @@ import 'package:openfoodfacts/personalized_search/preference_importance.dart';
 import 'package:openfoodfacts/personalized_search/product_preferences_manager.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/helpers/attributes_card_helper.dart';
-import 'package:smooth_app/pages/user_preferences_dev_mode.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_dev_mode.dart';
 
 /// Match and score of a [Product] vs. Preferences
 ///
