+++ packages/smooth_app/lib/pages/onboarding/preferences_page.dart
@@ -10,8 +10,8 @@ import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
 import 'package:smooth_app/pages/onboarding/next_button.dart';
 import 'package:smooth_app/pages/onboarding/onboarding_flow_navigator.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_food.dart';
 import 'package:smooth_app/pages/product/summary_card.dart';
-import 'package:smooth_app/pages/user_preferences_food.dart';
 
 class PreferencesPage extends StatefulWidget {
   const PreferencesPage(this._localDatabase) : super();
