+++ packages/smooth_app/lib/pages/preferences/abstract_collapsible_user_preferences.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
-import 'package:smooth_app/pages/abstract_user_preferences.dart';
+import 'package:smooth_app/pages/preferences/abstract_user_preferences.dart';
 
 /// Abstraction of a collapsed/expanded display for the preference pages.
 abstract class AbstractCollapsibleUserPreferences
