+++ packages/smooth_app/lib/pages/preferences/abstract_user_preferences.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/generic_lib/animations/smooth_animated_collapse_arrow.dart';
-import 'package:smooth_app/pages/user_preferences_list_tile.dart';
-import 'package:smooth_app/pages/user_preferences_page.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_list_tile.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_page.dart';
 import 'package:smooth_app/themes/constant_icons.dart';
 
 /// Abstraction of a display for the preference pages.
