+++ packages/smooth_app/lib/pages/preferences/user_preferences_page.dart
@@ -4,16 +4,16 @@ import 'package:provider/provider.dart';
 import 'package:smooth_app/data_models/product_preferences.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
-import 'package:smooth_app/pages/abstract_user_preferences.dart';
 import 'package:smooth_app/pages/all_user_product_list_page.dart';
-import 'package:smooth_app/pages/user_preferences_connect.dart';
-import 'package:smooth_app/pages/user_preferences_contribute.dart';
-import 'package:smooth_app/pages/user_preferences_dev_mode.dart';
-import 'package:smooth_app/pages/user_preferences_faq.dart';
-import 'package:smooth_app/pages/user_preferences_food.dart';
-import 'package:smooth_app/pages/user_preferences_list_tile.dart';
-import 'package:smooth_app/pages/user_preferences_profile.dart';
-import 'package:smooth_app/pages/user_preferences_settings.dart';
+import 'package:smooth_app/pages/preferences/abstract_user_preferences.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_connect.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_contribute.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_dev_mode.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_faq.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_food.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_list_tile.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_profile.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_settings.dart';
 import 'package:smooth_app/themes/constant_icons.dart';
 import 'package:smooth_app/themes/theme_provider.dart';
