+++ packages/smooth_app/lib/pages/preferences/user_preferences_profile.dart
@@ -10,10 +10,10 @@ import 'package:smooth_app/generic_lib/design_constants.dart';
 import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
 import 'package:smooth_app/helpers/analytics_helper.dart';
 import 'package:smooth_app/helpers/launch_url_helper.dart';
-import 'package:smooth_app/pages/abstract_user_preferences.dart';
 import 'package:smooth_app/pages/onboarding/country_selector.dart';
+import 'package:smooth_app/pages/preferences/abstract_user_preferences.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_page.dart';
 import 'package:smooth_app/pages/user_management/login_page.dart';
-import 'package:smooth_app/pages/user_preferences_page.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 /// Collapsed/expanded display of profile for the preferences page.
