+++ packages/smooth_app/lib/widgets/tab_navigator.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:smooth_app/pages/history_page.dart';
 import 'package:smooth_app/pages/page_manager.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_page.dart';
 import 'package:smooth_app/pages/scan/scan_page.dart';
-import 'package:smooth_app/pages/user_preferences_page.dart';
 
 class TabNavigator extends StatelessWidget {
   const TabNavigator({
