+++ packages/smooth_app/test/pages/user_preferences_page_test.dart
@@ -6,7 +6,7 @@ import 'package:openfoodfacts/personalized_search/product_preferences_selection.
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:smooth_app/data_models/product_preferences.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
-import 'package:smooth_app/pages/user_preferences_page.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_page.dart';
 import 'package:smooth_app/themes/theme_provider.dart';
 
 import '../utils/goldens.dart';
